requirements.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.2
django-rest-swagger==0.3.2
django-tastypie==0.12.1
djangorestframework==3.1.2
gunicorn==19.3.0
Pygments==2.0.2
PyMySQL==0.6.6
python-dateutil==2.4.2
python-mimeparse==0.1.4
PyYAML==3.11
six==1.9.0
South==1.0.2
static3==0.6.1
wheel==0.24.0
